+++ include/Mesh.h
@@ -11,9 +11,9 @@ class Mesh
 public:
 	Mesh(aiMesh* mesh);
 	std::vector<Vertex>& getVertices() noexcept;
-	std::vector<unsigned short> getIndices() noexcept;
+	std::vector<unsigned int> getIndices() noexcept;
 	void Transform(DirectX::FXMMATRIX matrix);
 private:
 	std::vector<Vertex> vertices;
-	std::vector<unsigned short> indices;
+	std::vector<unsigned int> indices;
 };
